Change robot position type to double

+++ cpp-console/robot.cpp
@@ -3,7 +3,8 @@
 
 #include <exception>
 
-Robot::Robot(int x, int y, World& world)
+//Robot::Robot(int x, int y, World& world)
+Robot::Robot(double x, double y, World& world)
 	: x(x >= 0 && x < world.getWidth() ? x : throw std::invalid_argument("Robot X coordinate is out of the world."))
 	, y(y >= 0 && y < world.getHeight() ? y : throw std::invalid_argument("Robot Y coordinate is out of the world."))
 	, world(world)
